core/requirements.txt
aiofiles>=0.7.0
alembic>=1.7.1
asyncpg>=0.27.0
autopep8>=2.3.2
black>=25.1.0
easyocr>=1.7.0
email-validator>=2.0.0
fastapi>=0.68.0
gotrue>=1.0.0
itsdangerous>=2.0.0
jinja2>=3.0.1
passlib[bcrypt]>=1.7.4
pillow>=10.0.0

# Development dependencies
pre-commit>=3.5.0
psycopg2-binary>=2.9.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pylint>=3.3.4
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
requests>=2.26.0
ruff>=0.9.6
slowapi>=0.1.4
sqlalchemy[asyncio]>=1.4.23
starlette>=0.14.2
supabase==2.13.0
tenacity>=8.0.1
urllib3<2.0.0
uvicorn[standard]>=0.15.0
